recommender_microservice_FaaS/requirements.txt
flask
pandas==1.0.5
numpy==1.19.1
sqlalchemy==1.3.18
sqlalchemy-utils
scikit-learn
h5py~=2.10.0
google-cloud-storage